Updated some configs (and added a note in commit message!)
Note : Latitude and Longitude value will be different at GCS even though the message is not spoofed. This is because ADS-B format only has 17bits to represent Latitude and Longitude.

By combining even/odd frames together, final position accuracy is typically around 5-10 meters.

So don't panic... It is working as it should be.